+++ Biomass.hpp
@@ -12,7 +12,7 @@ class abstract_Biomass {
 protected:
     int age_of_cell = 0;
     int max_count_reps = simulation_config::biomass::max_count_reps;
-    float biomass = simulation_config::biomass::initial_biomass;
+    double biomass = simulation_config::biomass::initial_biomass;
     Cell* nucleus = nullptr;
     void copy_common_state_to(abstract_Biomass& other) const;
 
@@ -24,7 +24,7 @@ public:
 
     int get_age() const;
     float get_level_of_resistance() const;
-    float get_biomass() const;
+    double get_biomass() const;
 
     // Теперь must_he_die принимает антибиотик
     bool must_he_die(Food& food, const Antibiotic& antibiotic) const;
@@ -37,7 +37,7 @@ public:
 
     virtual void consume_and_decay(Food& food);
     virtual float is_active_for_monod() const = 0;
-    virtual float get_maintenance_rate() const = 0;
+    virtual double get_maintenance_rate() const = 0;
 
     virtual void step_after_death();
     virtual bool should_be_removed_from_field() const;
@@ -54,7 +54,7 @@ public:
     bool reproduction(Field& current_field, int x, int y) override;
     void consume_and_decay(Food& food) override;
     float is_active_for_monod() const override { return 1.0f; }
-    float get_maintenance_rate() const override { return static_cast<float>(simulation_config::monod::m_act); }
+    double get_maintenance_rate() const override { return simulation_config::monod::m_act; }
 };
 
 class nonactive_Biomass : public abstract_Biomass {
@@ -73,7 +73,7 @@ public:
     bool reproduction(Field& current_field, int x, int y) override;
     
     float is_active_for_monod() const override { return 0.0f; }
-    float get_maintenance_rate() const override { return static_cast<float>(simulation_config::monod::m_inactiv); }
+    double get_maintenance_rate() const override { return simulation_config::monod::m_inactiv; }
 
     // Применяет коэффициенты дормантности (устойчивость/возраст/расход пищи)
     // к уже скопированному общему состоянию клетки. Вызывается сразу после
@@ -93,7 +93,7 @@ public:
     int count_of_steps_from_death = 0;
     
     float is_active_for_monod() const override { return 0.0f; }
-    float get_maintenance_rate() const override { return 0.0f; }
+    double get_maintenance_rate() const override { return 0.0; }
     
     void step_after_death() override;
     bool should_be_removed_from_field() const override;
